+++ examples/example_shred/src/test_payroll_calculation.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+
+if [[ "${CI:-}" == "" ]]; then
+  . ./global_variables.sh
+fi
+
 chmod +x ${EXECUTABLE_NAME}
 echo "1000.00" > test_input.dat
 echo "Running test with input from test_input.dat..."